Mypy and Flake8 are now executed with python -m when option to search in interpreter is used.

+++ plugins/com.python.pydev.analysis/src/com/python/pydev/analysis/flake8/Flake8Visitor.java
@@ -61,15 +61,18 @@ import com.python.pydev.analysis.external.IExternalCodeAnalysisStream;
             return;
         }
 
-        File flake8Location = Flake8Preferences.getFlake8Location(pythonNature);
-        if (flake8Location == null || !FileUtils.enhancedIsFile(flake8Location)) {
-            if (flake8Location == null) {
-                Log.log("Unable to find flake8. Project: " + project.getName());
-            } else {
-                Log.log("flake8 location does not exist: " + flake8Location);
+        File flake8Location = null;
+        if (!Flake8Preferences.useFlake8FromPythonNature(pythonNature, project)) {
+            flake8Location = Flake8Preferences.getFlake8Location(pythonNature);
+            if (flake8Location == null || !FileUtils.enhancedIsFile(flake8Location)) {
+                if (flake8Location == null) {
+                    Log.log("Unable to find flake8. Project: " + project.getName());
+                } else {
+                    Log.log("flake8 location does not exist: " + flake8Location);
+                }
+                deleteMarkers();
+                return;
             }
-            deleteMarkers();
-            return;
         }
 
         try {
